Restore original algorithm for clone detection
This algorithm described in paper, which we took as a base of our
research. Implementation of this algorithm was initially contributed
by Sharif Uddin, but was lost later, because was replaced on another
algorithm contributed by Andrew Tereskin. So this is a restoration
with some improvements like increased performance and more clean code.