add optimization with request cache in DbCloneIndex
optimization is based on fact that requests are costly, so decrease number of requests from number of blocks in analyzed project to number of files in analyzed project.  as all requests CloneIndex#getBySequenceHash usually are performed sequentially for blocks in file try run query for all blocks and same in HashMap.